# Add key and value methods to DebugMap