@GuillaumeGomez 2019-05-02
@rust-lang/infra: seems like a condition is broken:

> Some changes occurred in HTML/CSS.
> 
> cc @GuillaumeGomez