@pietroalbini 2019-05-02
Highfive is working as expected though:

https://github.com/rust-lang/highfive/blob/748c035ca2a913cb3bc95404f8665985736f165c/highfive/configs/rust-lang/rust.json#L50-L53