@KodrAus 2019-05-13
Very quick, very dirty benchmark:

```rust
#![feature(test)]

extern crate test;

use std::{
    cell::Cell,
    fmt::{self, Write},
};

struct NoOpWrite;

impl Write for NoOpWrite {
    fn write_str(&mut self, _: &str) -> fmt::Result {
        Ok(())
    }
}

#[bench]
fn debugmap_entry(b: &mut test::Bencher) {
    struct Map<'a>(Cell<Option<&'a mut test::Bencher>>);

    impl<'a> fmt::Debug for Map<'a> {
        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
            let mut map = f.debug_map();

            let b = self.0.take().unwrap();
            b.iter(|| {
                map.entry(&"a", &42);
            });

            map.finish()
        }
    }

    let _ = write!(NoOpWrite, "{:?}", Map(Cell::new(Some(b))));
}
```

On my machine, running this on the current `nightly` and this PR yields:

| Before | After |
| ------ | ----- |
| 36 ns/iter (+/- 7) | 44 ns/iter (+/- 2) |

So there's a bit of a performance impact like you'd expect, but I don't think it's too significant, and if it is we can inline the implementation in `entry`.